Refactor whenReady into shared utility
