+++ challenge/game_challenge.js
@@ -320,15 +320,9 @@ document.addEventListener("keydown", (e) => {
 });
 
 // --- Contrôles boutons physiques (optionnel)
-function whenReady(fn) {
-  if (window.cordova || window.Capacitor) {
-    document.addEventListener('deviceready', fn, false);
-  } else if (document.readyState !== 'loading') {
-    fn();
-  } else {
-    document.addEventListener('DOMContentLoaded', fn);
-  }
-}
+const { whenReady } = typeof module !== 'undefined' && module.exports
+  ? require('../scripts/utils')
+  : window;
 
 whenReady(() => {
   const btnLeft   = document.querySelector("button[data-action='left']");
